Update the saving design and behaviours
Why?
With the save button on the top of the language table, the user will save the form and be redirected to the Currency List Page without allowing him to edit the currency format. Because he could saw the table, it will be frustrating for him to not be able to edit it directly.
The old design was designed like this but was changed because it was confusing to have a save button on the pop in for the currency format and as well for saving the form.
To compensate the fact that the user has to press on the save button any time he made an update on currency format. We put the wording "Apply" instead of "Save" to explicit that there is only one real button to complete and save the form.